[#3]feat: UserController 생성 및 routes에 Controller 등록
